+++ app/src/main/java/eg/gov/iti/skyscanner/models/Repository.kt
@@ -21,9 +21,10 @@ class Repository private constructor(
         }
     }
 
+
     override suspend fun getRetrofitWeather(
-        lat: Float,
-        lon: Float,
+        lat: Double,
+        lon: Double,
         units: String,
         lang: String,
         apiKey: String
