+++ app/src/main/java/eg/gov/iti/skyscanner/models/RepositoryInterface.kt
@@ -3,6 +3,6 @@ package eg.gov.iti.skyscanner.models
 import kotlinx.coroutines.flow.Flow
 
 interface RepositoryInterface {
-    suspend fun getRetrofitWeather(lat:Float,lon:Float,units:String,lang:String,apiKey:String): Flow<WeatherDetail>?
+    suspend fun getRetrofitWeather(lat:Double,lon:Double,units:String,lang:String,apiKey:String): Flow<WeatherDetail>?
 
 }
